+++ GHEtool/VariableClasses/__init__.py
@@ -6,3 +6,4 @@ from .CustomGFunction import CustomGFunction, load_custom_gfunction, _timeValues
 from .GFunction import GFunction, FIFO
 from .SizingSetup import SizingSetup
 from .Borehole import Borehole
+from .TemperatureLimits import *